requirements.txt
pytest
playwright
pytest-playwright
allure-pytest
python-dotenv
requests
pydantic